feat: updated codebase structure
