+++ front/myapp-front/src/components/actionModal/alertDialog/delete/DeleteAlertDialog.types.ts
@@ -1,0 +1,3 @@
+export type DeleteAlertDialogProps = {
+  elementId: string;
+};
