+++ front/myapp-front/src/components/actionModal/group/ModalGroup.types.ts
@@ -1,0 +1,6 @@
+import { GroupDetails } from "../../../types/interfaces";
+import { BaseActionModalProps } from "../ActionModal.types";
+
+export type GroupModalProps = {
+  initialData: GroupDetails;
+} & Omit<BaseActionModalProps, "btnName">;
