+++ front/myapp-front/src/components/actionModal/header/ActionModalHeader.types.ts
@@ -1,0 +1,9 @@
+import { DeleteAlertDialogProps } from "../alertDialog/delete/DeleteAlertDialog.types";
+
+export type ActionModalHeaderProps = {
+  title: string;
+  isShow: boolean;
+  setFormAction: React.Dispatch<
+    React.SetStateAction<"show" | "create" | "edit">
+  >;
+} & DeleteAlertDialogProps;
