+++ front/myapp-front/src/components/actionModal/invitation/ModalInvitation.types.ts
@@ -1,0 +1,6 @@
+import { InvitationDetails } from "../../../types/interfaces";
+import { BaseActionModalProps } from "../ActionModal.types";
+
+export type InvitationModalProps = {
+  initialData: InvitationDetails;
+} & Omit<BaseActionModalProps, "btnName">;
