+++ front/myapp-front/src/components/actionModal/tag/ModalTag.types.ts
@@ -1,0 +1,7 @@
+import { TagFormDetails } from "../../../types/interfaces";
+import { BaseActionModalProps } from "../ActionModal.types";
+
+export type ModalTagProps = {
+  initialData: TagFormDetails;
+  handleSubmit: (data: TagFormDetails) => void;
+} & Omit<BaseActionModalProps, "btnName">;
