+++ front/myapp-front/src/components/actionModal/task/ModalTask.types.ts
@@ -1,0 +1,7 @@
+import { TaskFormDetails } from "../../../types/interfaces";
+import { BaseActionModalProps } from "../ActionModal.types";
+
+export type ModalTaskProps = {
+  initialData: TaskFormDetails;
+  handleSubmit: (data: TaskFormDetails) => void;
+} & Omit<BaseActionModalProps, "btnName">;
