+++ front/myapp-front/src/components/card/invitation/InvitationCard.types.ts
@@ -1,0 +1,5 @@
+import { InvitationDetails } from "../../../types/interfaces";
+
+export type InvitationRendererProps = {
+  element: InvitationDetails;
+};
