+++ front/myapp-front/src/components/card/tag/TagCard.types.ts
@@ -1,0 +1,5 @@
+import { TagDetails } from "../../../types/interfaces";
+
+export type TagRendererProps = {
+  element: TagDetails;
+};
