+++ front/myapp-front/src/components/card/task/TaskCard.types.ts
@@ -1,0 +1,5 @@
+import { DividedTaskDetails } from "../../../types/interfaces";
+
+export type TaskRendererProps = {
+  element: DividedTaskDetails;
+};
