+++ front/myapp-front/src/components/card/user/UserCard.tsx
@@ -1,10 +1,10 @@
 import React, { useContext } from "react";
-import { UserRendererProps } from "./Card.types";
 import { Avatar, IconButton, Paper, Stack, Typography } from "@mui/material";
 import DeleteIcon from "@mui/icons-material/Delete";
-import { GroupContext } from "../../context/group/GroupContext";
-import { fetchIconUrl } from "../../utils/fetchUserIcon";
-import { UserContext } from "../../context/user/UserContext";
+import { GroupContext } from "../../../context/group/GroupContext";
+import { fetchIconUrl } from "../../../utils/fetchUserIcon";
+import { UserContext } from "../../../context/user/UserContext";
+import { UserRendererProps } from "./UserCard.types";
 
 function UserCard({ element }: UserRendererProps) {
   const { state: userState, dispatch: userDispatch } = useContext(UserContext);
