+++ front/myapp-front/src/components/card/user/UserCard.types.ts
@@ -1,0 +1,5 @@
+import { DividedUserDetails } from "../../../types/interfaces";
+
+export type UserRendererProps = {
+  element: DividedUserDetails;
+};
